replace paths for moved files
