+++ packages/modules/create-itwin-viewer/templates/desktop/src/backend/main.ts
@@ -19,8 +19,8 @@ import { Menu, shell } from "electron";
 import type { MenuItemConstructorOptions } from "electron/main";
 import * as path from "path";
 
-import { AppLoggerCategory } from "../common/LoggerCategory";
-import { channelName } from "../common/types";
+import { AppLoggerCategory } from "../types";
+import { channelName } from "../types";
 import ViewerHandler from "./ViewerHandler";
 
 require("dotenv-flow").config(); // eslint-disable-line @typescript-eslint/no-var-requires
