+++ packages/modules/create-itwin-viewer/templates/desktop/tsconfig.backend.json
@@ -23,7 +23,6 @@
     "outDir": "./lib"
   },
   "include": [
-    "src/backend/*.ts",
-    "src/common/*.ts"
+    "src/backend/*.ts"
   ]
 }
